offload dynamic vector into individual files
